% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAC_PSSM.R
\name{aac_pssm}
\alias{aac_pssm}
\title{AAC-PSSM feature vectors}
\usage{
aac_pssm(pssm_name)
}
\arguments{
\item{pssm_name}{name of PSSM Matrix file}
}
\value{
feature vector of length 20
}
\description{
AAC-PSSM Feature stands for Amino Acid composition which is actually mean of PSSM Matrix
columns which its length is 20. combination of this feature vector and DPC-PSSM feature vector would be
AADP-PSSM feature vector.
}
\examples{
X<-aac_pssm(system.file("extdata", "C7GQS7.txt.pssm", package="PSSMCOOL"))
}
\references{
Liu, T., Zheng, X. and Wang, J. (2010) Prediction of protein structural class for
low-similarity sequences using support vector machine and PSI-BLAST profile, Biochimie, 92, 1330-1334.
}
