% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSSM_BLOCK.R
\name{Averag_Block}
\alias{Averag_Block}
\title{Averag Block feature vector}
\usage{
Averag_Block(pssm_name)
}
\arguments{
\item{pssm_name}{name of PSSM Matrix file}
}
\value{
feature vector of length 400
}
\description{
In this feature at first PSSM Matrix is divided to 20 Blocks. Then for each
Block mean of columns is computed to get 20-dimensional vector, eventually by appending these vectors to each other final feature
vector is obtained which would be of length 400. this feature vector is similar to \code{\link{PSSMBLOCK}} for N=20.
}
\examples{
v<-Averag_Block(system.file("extdata", "C7GQS7.txt.pssm", package="PSSMCOOL"))
}
\references{
L. Nanni, A. Lumini, and S. J. T. S. W. J. Brahnam, "An empirical study of different approaches for protein classification,"
vol. 2014, 2014.
}
