% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DPC_PSSM.R
\name{dpc_pssm}
\alias{dpc_pssm}
\title{DPC_PSSM feature vector}
\usage{
dpc_pssm(pssm_name)
}
\arguments{
\item{pssm_name}{name of PSSM Matrix file}
}
\value{
feature vector of length 400
}
\description{
This Feature stands for dipeptide composition, to get this feature vector for both different
columns the elements in two consecutive rows corresponds to this columns, would be multiplied and this scenario
is done  for all L-1 consecutive rows (L is protein length) and finally summation of these numbers divides by
L-1. since the result depends on two different columns, eventually a Feature vector of length 400
will be obtained.
}
\examples{
X<-dpc_pssm(system.file("extdata", "C7GQS7.txt.pssm", package="PSSMCOOL"))

}
\references{
Liu, T., Zheng, X. and Wang, J. (2010) Prediction of protein structural class for
low-similarity sequences using support vector machine and PSI-BLAST profile, Biochimie, 92, 1330-1334.
}
