% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCSH.R
\name{k_mers}
\alias{k_mers}
\title{3-mer and 2-mer in dataframe}
\usage{
k_mers(s, h)
}
\arguments{
\item{s}{a dataframe with 2 columns}

\item{h}{is length of k-mer}
}
\value{
all k-mers by counting paths of length h in dataframe which is considered as a graph
}
\description{
This function produces all possible 2-mers or 3-mers by counting paths of length 2 or 3 in
a dataframe which is thought as a graph
}
\examples{
s1<-LETTERS[1:4]
s2<-LETTERS[3:6]
s<-data.frame(s1,s2)
dc<-k_mers(s,3)
}
