% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_seperated.R
\name{k_separated_bigrams_pssm}
\alias{k_separated_bigrams_pssm}
\title{k_separated_bigrams_pssm feature vector}
\usage{
k_separated_bigrams_pssm(pssm_name, k = 1)
}
\arguments{
\item{pssm_name}{is name of PSSM Matrix file}

\item{k}{a parameter that specifies separated length between amino acids}
}
\value{
a feature vector of length 400
}
\description{
This feature is almost identical to the \code{\link{dpc_pssm}} feature, and in fact the DPC feature is part of
this feature (for k=1) and for two different columns, considers rows that differ by the size of the unit k.
}
\examples{
 X<-k_separated_bigrams_pssm(system.file("extdata", "C7GQS7.txt.pssm", package="PSSMCOOL"),1)
}
\references{
Saini, H., et al.(2016) Protein Fold Recognition Using Genetic Algorithm Optimized Voting Scheme and Profile
Bigram.
}
