% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_seperated.R
\name{k_seperated_bigrame}
\alias{k_seperated_bigrame}
\title{k_seperated_bigrame feature vector}
\usage{
k_seperated_bigrame(pssm_name, k)
}
\arguments{
\item{pssm_name}{is name of PSSM Matrix file}

\item{k}{a parameter that specifies separated length between amino acids}
}
\value{
a feature vector of length 400
}
\description{
This feature is almost identical to the \code{\link{DPC_PSSM}} feature, and in fact the DPC feature is part of
this feature (for k=1) and for two different columns, considers rows that differ by the size of the unit k.
}
\examples{
 w<-k_seperated_bigrame(system.file("extdata", "C7GQS7.txt.pssm", package="PSSMCOOL"),5)
}
\references{
Saini, H., et al.(2016) Protein Fold Recognition Using Genetic Algorithm Optimized Voting Scheme and Profile
Bigram.
}
