% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSSM_BLOCK.R
\name{PSSMBLOCK}
\alias{PSSMBLOCK}
\title{PSSM BLOCK feature vector}
\usage{
PSSMBLOCK(pssm_name, N)
}
\arguments{
\item{pssm_name}{neme of PSSM Matrix file}

\item{N}{number of blocks}
}
\value{
feature vector that it's length depends on parameter N
}
\description{
In this feature at first PSSM Matrix is divided to Blocks based on Number N which user imports. Then for each
Block mean of columns is computed to get 20-dimensional vector, eventually by appending these vectors to each other final feature
vector is obtained.
}
\examples{
as<-PSSMBLOCK(system.file("extdata", "C7GQS7.txt.pssm", package="PSSMCOOL"),5)
}
\references{
J.-Y. An, L. Zhang, Y. Zhou, Y.-J. Zhao, and D.-F. J. J. o. c. Wang, "Computational methods using weighed-extreme learning
machine to predict protein self-interactions with protein evolutionary information," vol. 9, no. 1, p. 47, 2017.
}
