% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCSH.R
\name{scsh2}
\alias{scsh2}
\title{SCSH Feature vector}
\usage{
scsh2(pssm_name, k)
}
\arguments{
\item{pssm_name}{is name of PSSM Matrix file}

\item{k}{a parameter indicates length of k-mer}
}
\value{
feature vector of length 400 or 8000
}
\description{
This function gets a PSSM Matrix as input and extracts corresponding protein and consunsus
sequence from it. By placing these two vectors next to each other, a dataframe is created.
In each row of this dataframe, each component is connected to the next row components by an arrow. so a
directed graph is produced. next by use of previous functions a feature vector of length 400 or 8000 is created.
}
\examples{
zz<- scsh2(system.file("extdata","C7GRQ3.txt.pssm",package="PSSMCOOL"),2)
}
\references{
Zahiri, J., et al., LocFuse: human protein–protein interaction prediction via classifier fusion using protein
localization information. Genomics, 2014. 104(6): p. 496-503.
}
