% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_average.R
\name{single_Average}
\alias{single_Average}
\title{single Average feature}
\usage{
single_Average(pssm_name)
}
\arguments{
\item{pssm_name}{name of PSSM Matrix file}
}
\value{
feature vector of length 400
}
\description{
This descriptor is a variant of the Average Block descriptor and is designed to group together rows
related to the same amino acid, thus considering domains of a sequence with similar conservation rates.
}
\examples{
w<-single_Average(system.file("extdata", "C7GQS7.txt.pssm", package="PSSMCOOL"))
}
\references{
L. Nanni, A. Lumini, and S. J. T. S. W. J. Brahnam, "An empirical study of different approaches for protein classification,"
vol. 2014, 2014.
}
\seealso{
\code{\link{Averag_Block}}
}
