% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/somaPssm.R
\name{SOMA_PSSM}
\alias{SOMA_PSSM}
\title{SOMA PSSM Feature}
\usage{
SOMA_PSSM(pssm_name)
}
\arguments{
\item{pssm_name}{name of PSSM file}
}
\value{
feature vector of length 160
}
\description{
In this function each column can be viewed as a stochastic time series, and each PSSM contains 20
columns, in other words, each PSSM contains 20 stochastic time series and Second-order moving average (SOMA)
algorithm is applied to these columns to extract SOMA PSSM feature vector.
}
\examples{
w<-SOMA_PSSM(system.file("extdata", "C7GQS7.txt.pssm", package="PSSMCOOL"))
}
\references{
Y. Liang, S. J. J. o. M. G. Zhang, and Modelling, "Predict protein structural class by incorporating two different modes of evolutionary
information into Chou's general pseudo amino acid composition," vol. 78, pp. 110-117, 2017.
}
