% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCSH.R
\name{three_mer}
\alias{three_mer}
\title{3-Mer and 2-Mer}
\usage{
three_mer(k)
}
\arguments{
\item{k}{is length of k-mer which user imports}
}
\value{
a matrix which its first row includes all k-mers
}
\description{
This function produces all possible k-mers from 20 amino acids for use in other functions.
}
\examples{
ax<-three_mer(3)
}
