% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAtrigrame_pssm.R
\name{trigrame_pssm}
\alias{trigrame_pssm}
\title{trigrame feature vector}
\usage{
trigrame_pssm(pssm_name)
}
\arguments{
\item{pssm_name}{name of PSSM Matrix file}
}
\value{
feature vector of lenght 8000
}
\description{
This feature vector is 8000-dimentional feature vector wich is computed from  tri-gram probability matrix
T obtained from PSSM Matrix.to achieve this purpose elements in three successive rows and arbitrary columns  are multiplied
together then these results are added together by changing variable i from 1 to L-1, which i is counter of row and L
indicates protein length. since there are 20 columns thus final feature vector would be of length 8000.
}
\examples{
as<-trigrame_pssm(paste0(system.file("extdata",package="PSSMCOOL"),"/C7GSI6.txt.pssm"))
}
\references{
Paliwal, K.K., et al. (2014) A tri-gram based feature extraction technique using linear probabilities of position
specific scoring matrix for protein fold recognition, IEEE transactions on nanobioscience, 13, 44-50
}
