\name{PSSMCOOL-package}

\alias{PSSMCOOL-package}

\title{PSSMCOOL:Extracting Various Feature vectors from PSSM Matrix}

\docType{package}

\description{
  This package is a Comprehensive toolkit for generating various numerical features from PSSM Matrix correspond to each protein sequence and covers all features introduced in POSSUM Package and some other new features. see \href{https://rpubs.com/alireza_69/640294}{PSSMCOOL_Rpubs} for more information.
  }

\seealso{

\strong{All functions in this package:}
\itemize{
  \item \code{\link{AATP_TPCC}}
  \item \code{\link{AB_PSSM}}
  \item \code{\link{Averag_Block}}
  \item \code{\link{consunsus_sequence}}
  \item \code{\link{CS_PSe_PSSM}}
  \item \code{\link{DFMCA_PSSM}}
  \item \code{\link{Discrete_Cosine_Transform}}
  \item \code{\link{disulfid}}
  \item \code{\link{DP_PSSM}}
  \item \code{\link{DPC_PSSM}}
  \item \code{\link{dwt_PSSM}}
  \item \code{\link{EDP_MEDP}}
  \item \code{\link{FPSSM}}
  \item \code{\link{FPSSM2}}
  \item \code{\link{grey_pssm_pseAAC}}
  \item \code{\link{k_mers}}
  \item \code{\link{k_seperated_bigrame}}
  \item \code{\link{kiderafactor}}
  \item \code{\link{LPC_PSSM}}
  \item \code{\link{MBMGACPSSM}}
  \item \code{\link{pse_pssm}}
  \item \code{\link{pssm400}}
  \item \code{\link{pssm_ac}}
  \item \code{\link{pssm_cc}}
  \item \code{\link{PSSM_SD}}
  \item \code{\link{pssm_seg}}
  \item \code{\link{PSSMAC}}
  \item \code{\link{PSSMBLOCK}}
  \item \code{\link{rpssm}}
  \item \code{\link{scsh2}}
  \item \code{\link{single_Average}}
  \item \code{\link{smoothed_PSSM}}
  \item \code{\link{SOMA_PSSM}}
  \item \code{\link{SVD_PSSM}}
  \item \code{\link{trigrame_pssm}}
}
\strong{Useful links:}
\itemize{
  \item \url{https://rpubs.com/alireza_69/640294}
  \item \url{https://github.com/Alireza9651501005/PSSMCOOL/}
  \item \url{http://bioinf.modares.ac.ir/software/pssm/DownloadALL.html}
  \item Report bugs at \url{https://github.com/Alireza9651501005/PSSMCOOL/issues}
  }
}
\author{
\strong{Maintainer}: Alireza Mohammadi \email{alireza691111@gmail.com}

Authors:
\itemize{
  \item Alireza Mohammadi \email{alireza691111@gmail.com}
  \item Javad Zahiri \email{zahiri@modares.ac.ir}
}

}

